copy to clipboard
This branch explores the inclusion of a copy to clipboard button on the code chunks. It kind of requires a change to the highlight (kate) because that adds sourceCode labels to the source code chunks which is the hook that is used to identify where the copy button should be included.